Refactor error components into individual files
