+++ ui/__tests__/components/error.test.js
@@ -1,8 +1,7 @@
 import { shallow } from 'enzyme';
 import React from 'react';
 
-import ErrorView, { FourOhFour, UnexpectedError, UserError } from '../../components/error';
-
+import ErrorView from '../../components/error';
 
 describe('<ErrorView />', () => {
   it('renders a 404', () => {
@@ -20,45 +19,5 @@ describe('<ErrorView />', () => {
     const result = shallow(<ErrorView statusCode={1234} />);
     expect(result.name()).toBe('UnexpectedError');
     expect(result.prop('statusCode')).toBe(1234);
-  });
-});
-
-describe('<FourOhFour />', () => {
-  it('has expected milestones', () => {
-    const wrapped = shallow(<FourOhFour />);
-    expect(wrapped.name()).toBe('HeaderFooter');
-    const text = wrapped.childAt(0).text();
-    expect(text).toMatch(/We can’t find the page/); // note the smart quote
-    expect(text).toMatch(/contact us/);
-    expect(text).toMatch(/Return home/);
-  });
-});
-
-describe('<UserError />', () => {
-  it('has expected milestones', () => {
-    const wrapped = shallow(<UserError message="This is my message" />);
-    expect(wrapped.name()).toBe('HeaderFooter');
-    const text = wrapped.childAt(0).text();
-    expect(text).toMatch(/Invalid Request/);
-    expect(text).toMatch(/This is my message/);
-    expect(text).toMatch(/Return home/);
-  });
-});
-
-describe('<UnexpectedError />', () => {
-  it('has the correct status code', () => {
-    const text = shallow(<UnexpectedError statusCode={502} />).text();
-    expect(text).toMatch(/Bad Gateway/);
-    expect(text).not.toMatch(/Unexpected error/);
-    expect(text).toMatch(/homepage/);
-    expect(text).toMatch(/contact us/);
-  });
-
-  it('works without a status code', () => {
-    const text = shallow(<UnexpectedError />).text();
-    expect(text).not.toMatch(/Bad Gateway/);
-    expect(text).toMatch(/Unexpected error/);
-    expect(text).toMatch(/homepage/);
-    expect(text).toMatch(/contact us/);
   });
 });
